name fix and uneccesary file removed
