+++ src/main/java/usecases/delete_message/DeleteInputModel.java
@@ -1,6 +1,8 @@
-package usecases.message_manipulation.edit_message;
+package usecases.delete_message;
 
-public class EditInputModel {
+
+public class DeleteInputModel
+{
 
     int currentUserUid;
     int senderUid;
@@ -8,10 +10,10 @@ public class EditInputModel {
     String content;
     String time;
 
-    public EditInputModel(int senderUid, int chatUid, String content, String time, int currentUserUid)
+    public DeleteInputModel(int senderUid, int chatUid, String content, String time, int currentUserUid)
     {
         this.senderUid = senderUid;
-        this.currentUserUid =currentUserUid;
+        this.currentUserUid = currentUserUid;
         this.chatUid = chatUid;
         this.time = time;
         this.content = content;
