+++ src/main/java/usecases/delete_message/DeleteInteractor.java
@@ -1,7 +1,7 @@
-package usecases.message_manipulation.delete_message;
+package usecases.delete_message;
 
-import java.util.List;
 import java.util.ArrayList;
+import java.util.List;
 
 public class DeleteInteractor implements DeleteInputBoundary {
