+++ src/main/java/usecases/delete_message/DeleteOutputModel.java
@@ -1,4 +1,4 @@
-package usecases.message_manipulation.delete_message;
+package usecases.delete_message;
 
 import java.util.List;
