+++ src/main/java/usecases/edit_message/EditDSGateway.java
@@ -1,7 +1,7 @@
-package usecases.message_manipulation.edit_message;
+package usecases.edit_message;
 
-import java.util.List;
 import java.util.ArrayList;
+import java.util.List;
 
 public interface EditDSGateway {
     /**
