+++ src/main/java/usecases/edit_message/EditInputModel.java
@@ -1,8 +1,6 @@
-package usecases.message_manipulation.delete_message;
+package usecases.edit_message;
 
-
-public class DeleteInputModel
-{
+public class EditInputModel {
 
     int currentUserUid;
     int senderUid;
@@ -10,10 +8,10 @@ public class DeleteInputModel
     String content;
     String time;
 
-    public DeleteInputModel(int senderUid, int chatUid, String content, String time, int currentUserUid)
+    public EditInputModel(int senderUid, int chatUid, String content, String time, int currentUserUid)
     {
         this.senderUid = senderUid;
-        this.currentUserUid = currentUserUid;
+        this.currentUserUid =currentUserUid;
         this.chatUid = chatUid;
         this.time = time;
         this.content = content;
