+++ src/main/java/usecases/edit_message/EditInteractor.java
@@ -1,9 +1,9 @@
-package usecases.message_manipulation.edit_message;
+package usecases.edit_message;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class EditInteractor implements EditInputBoundary{
+public class EditInteractor implements EditInputBoundary {
 
     final EditOutputBoundary output;
     final EditDSGateway dataBase;
