+++ src/main/java/usecases/edit_message/EditOutputBoundary.java
@@ -1,4 +1,4 @@
-package usecases.message_manipulation.edit_message;
+package usecases.edit_message;
 
 public interface EditOutputBoundary
 {
