+++ src/main/java/usecases/edit_message/EditOutputModel.java
@@ -1,4 +1,4 @@
-package usecases.message_manipulation.edit_message;
+package usecases.edit_message;
 
 import java.util.List;
